Move build cli to its own file
